+++ cmd/geth/usage.go
@@ -149,9 +149,7 @@ var AppHelpFlagGroups = []flagGroup{
 	{
 		Name: "VIRTUAL MACHINE",
 		Flags: []cli.Flag{
-			utils.VMEnableJitFlag,
-			utils.VMForceJitFlag,
-			utils.VMJitCacheFlag,
+			utils.VMCacheFlag,
 		},
 	},
 	{
